Move connector tests into connectors crate
